% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_classic_pptx.R
\name{eha_classic_pptx}
\alias{eha_classic_pptx}
\title{Classic EHA powerpoint template}
\usage{
eha_classic_pptx(
  aspect = "16x9",
  toc = FALSE,
  toc_depth = 2,
  fig_width = 9.4,
  fig_height = 4.24,
  fig_caption = TRUE,
  df_print = "default",
  keep_md = FALSE,
  smart = TRUE,
  md_extensions = NULL,
  slide_level = NULL,
  pandoc_args = NULL
)
}
\arguments{
\item{aspect}{String. Aspect ratio for power point slides.
Either "16x9" or "4x3". Default is "16x9"}

\item{toc}{Logical. Should the slides have a table of contents?
Default is FALSE}

\item{toc_depth}{Numeric. How many levels should the table of contents have.
Default is 2.}

\item{fig_width}{Numeric. Width of figures. Default is 9.4}

\item{fig_height}{Numeric. Height of figures. Default is 4.24}

\item{fig_caption}{Logical. Should figures have captions? Default is TRUE}

\item{df_print}{String. Print data frames? Default is "default"}

\item{keep_md}{Logical. Should markdown document be kept? Default is FALSE}

\item{smart}{Logical. Produce typographically correct outs. Default is TRUE}

\item{md_extensions}{String. Additional \href{https://pandoc.org/MANUAL.html#pandocs-markdown}{markdown extensions}.
Default is NULL}

\item{slide_level}{Numeric. Defines the heading level that defines individual
slides. Default is NULL}

\item{pandoc_args}{Arguments to pass to pandoc.}
}
\description{
Styles for eha rmarkdown ppt. Used with a template to create presentations
that dynamically update with your data.
}
\seealso{
\code{\link[rmarkdown]{powerpoint_presentation}}, \code{\link[officedown]{rpptx_document}}
}
