% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_classic_xar.R
\name{eha_classic_xar}
\alias{eha_classic_xar}
\title{EHA Xaringan Theme}
\usage{
eha_classic_xar(aspect = "16x9")
}
\arguments{
\item{aspect}{String. Aspect ratio for slides.
Either "16x9" or "4x3". Default is "16x9"}
}
\value{
xaringan theme
}
\description{
A function to produce an EHA Xaringan theme with the classic font. This
function leans on \code{\link[xaringanthemer]{style_mono_light}} for many of
the theme arguments. It uses \code{\link[xaringan]{moon_reader}} to create
the markdown renderable object. Unlike \code{\link{eha_classic_pptx}}, the
plot outputs for html slides are not editable svgs.
}
