% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_flexdashboard.R
\name{eha_flex_dashboard}
\alias{eha_flex_dashboard}
\title{EcoHealth Alliance Flex Dashboard Theme}
\usage{
eha_flex_dashboard(
  orientation = "row",
  bg = "#224A55",
  fg = "#5EB9D6",
  primary = "#97C83E",
  secondary = "#B0A28A",
  fontUrl,
  ...
)
}
\arguments{
\item{orientation}{Determines whether level 2 headings are treated as dashboard rows or dashboard columns.}

\item{bg}{Background color for \code{bs_theme}}

\item{fg}{Foreground color for \code{bs_theme}}

\item{primary}{Primary color for \code{bs_theme}}

\item{secondary}{Secondary color for code \code{bs_theme}}

\item{fontUrl}{URL where font is hosted}

\item{...}{additional arguments to be passed to \code{flex_dashboard}}
}
\value{
function
}
\description{
Function to create a custom flexdashboard theme for EHA.
}
\seealso{
\code{\link[flexdashboard:flex_dashboard]{flexdashboard::flex_dashboard()}}
\code{\link[bslib:bs_theme]{bslib::bs_theme()}}
}
