% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehaletter.R
\name{eha_letter}
\alias{eha_letter}
\title{Create a letter on EHA letterhead}
\usage{
eha_letter(...)
}
\description{
This is a thin wrapper around the \code{\link[linl:linl]{linl::linl()}} function which prints letters,
automatically using the EHA letterhead.  See that function's options for
further
}
