% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_pptx_template.R
\name{eha_pptx_template}
\alias{eha_pptx_template}
\title{Open EHA Powerpoint templte}
\usage{
eha_pptx_template(
  template = c("classic", "avenir"),
  aspect = c("16x9", "4x3"),
  outdir = "./"
)
}
\arguments{
\item{template}{String. Either classic or avenir. Defines the template to be use.}

\item{aspect}{String. Either 16x9 or 4x3. Defines the aspect ratio of the slides}

\item{outfile}{String. Name of output file}
}
\value{

}
\description{
Open avenir and classic powerpoint templates. Providing no arguments will open
all four templates. Providing either an aspect or a template will return the
two corresponding templates. For example, if classic is provided to template,
then classic 16x9 and classic 4x3 will be opened.
}
\examples{
#open all the templates
eha_pptx_template()
}
