% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_xaringan.R
\name{eha_xaringan}
\alias{eha_xaringan}
\title{EHA Xaringan Theme}
\usage{
eha_xaringan(...)
}
\arguments{
\item{...}{not used}
}
\value{
xaringan theme
}
\description{
A function to produce an EHA Xaringan theme.
}
