% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehastyle-package.R
\docType{package}
\name{ehastyle-package}
\alias{ehastyle}
\alias{ehastyle-package}
\title{ehastyle: Default styles for EcoHealth Alliance Reports}
\description{
Contains R Markdown templates and ggplot2 themes suitable for EHA work
}
\author{
\strong{Maintainer}: Noam Ross \email{ross@ecohealthalliance.org}

Authors:
\itemize{
  \item Emma Mendelsohn \email{mendelsohn@ecohealthalliance.org}
  \item Collin Schwantes \email{schwantes@ecohealthalliance.org}
}

Other contributors:
\itemize{
  \item EcoHealth Alliance [copyright holder]
}

}
\keyword{internal}
