% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello function}
\usage{
hello(...)
}
\arguments{
\item{...}{additional arguments to pass to hello}
}
\value{
string
}
\description{
For testing markdown outputs
}
\examples{

hello()

}
