% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{outbreak}
\alias{outbreak}
\title{R Markdown format for EHA Outbreak Scenarios Reports}
\usage{
outbreak(
  self_contained = TRUE,
  lib_dir = NULL,
  keep_md = FALSE,
  cache_prefix = "cache/",
  ...
)
}
\arguments{
\item{self_contained}{Include all dependencies}

\item{lib_dir}{Local directory to copy assets}

\item{keep_md}{Keep knitr-generated markdown}

\item{cache_prefix}{String. Prefix for cache}

\item{...}{Additional function arguments to pass to the base R Markdown HTML
output formatter}

\item{theme}{Bootstrap theme}

\item{mathjax}{Include mathjax, "local" or "default"}

\item{pandoc_args}{Other arguments to pass to pandoc}
}
\description{
R Markdown format for EHA Outbreak Scenarios Reports
}
