% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{outbreak_word}
\alias{outbreak_word}
\title{Create output for Word document}
\usage{
outbreak_word(
  cache_prefix = "cache/",
  keep_md = FALSE,
  proposal = FALSE,
  usaid = FALSE,
  usaid_panel = 1,
  ...
)
}
\arguments{
\item{cache_prefix}{String. Prefix for cache}

\item{keep_md}{Logical. Keep the markdown file generated by knitting. Note that if this is TRUE then clean_supporting will always be FALSE.}

\item{proposal}{Logical. Should the proposal format be used.}

\item{usaid}{Logical. Should the USAID format be used}

\item{usaid_panel}{Numeric. Which USAID sidebar image should be used (1,2, or 4)?}

\item{...}{Additional arguments to pass to \code{\link[rmarkdown]{word_document}}}
}
\description{
Create output for Word document
}
\seealso{
\code{\link[rmarkdown]{knitr_options}}, \code{\link[rmarkdown]{output_format}}, \code{\link[rmarkdown]{word_document}}
}
