% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_dubug_functions.R
\name{ph_duplicate_check}
\alias{ph_duplicate_check}
\title{Check PPT template for duplicates}
\usage{
ph_duplicate_check(ppt_template, print = FALSE)
}
\arguments{
\item{ppt_template}{String. Path to ppt template}

\item{print}{Logical. Should the dataframe be printed?}
}
\value{
dataframe
}
\description{
Checks for duplicate PH elements that would break rendering.
}
\seealso{
\code{\link[officer]{read_pptx}}, \code{\link[officer]{layout_properties}}
}
