% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_resource.R
\name{pkg_resource}
\alias{pkg_resource}
\title{Get package resources}
\usage{
pkg_resource(...)
}
\arguments{
\item{...}{arguments to pass to \code{system.file}}
}
\value{
string
}
\description{
This function allows you to easily source package resources
like images or style sheets. The function will return a file
path to a resource stored in your package.
}
\examples{

logoEHA <- pkg_resource("eha_icon_logo.png")
print(logoEHA)
}
