% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_debug_functions.R
\name{ppt_layout}
\alias{ppt_layout}
\title{Print Powerpoint Layout}
\usage{
ppt_layout(ppt_template, print = FALSE)
}
\arguments{
\item{ppt_template}{String. Path to ppt file}

\item{print}{Logical. Should layout be printed}
}
\value{
data.frame
}
\description{
Convenience wrapper for\code{\link[officer]{layout_summary}} used to debug
issues with powerpoint templates. Returns a list with first element containing
the master names and the second element containing layout and master names for
a given ppt.
}
