% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakerLocations.R
\docType{data}
\name{speakerLocations}
\alias{speakerLocations}
\title{Synthetic speaker locations for airtables example.}
\format{
A data frame with 8 rows and 6 variables:
\describe{
\item{affiliation}{institution associated with speaker}
\item{cities}{location of institution}
\item{lat}{latitude of institution}
\item{lng}{longitude of institution}
\item{totalTalks}{number of talks from that institution}
\item{label}{summary of institution statistics}
...
}
}
\source{
airtable app id appwlxIzmQx5njRtQ as of 24 June 2021
}
\usage{
speakerLocations
}
\description{
Dataset containing speaker locations and labels.
}
\keyword{datasets}
