% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_flexdashboard.R
\name{eha_flex_dashboard}
\alias{eha_flex_dashboard}
\title{EcoHealth Alliance Flex Dashboard Theme}
\usage{
eha_flex_dashboard(
  orientation = "rows",
  bg = "#224A55",
  fg = "#5EB9D6",
  primary = "#97C83E",
  secondary = "#B0A28A",
  version = 4,
  fontUrl = NULL,
  family,
  style,
  weight,
  ...
)
}
\arguments{
\item{orientation}{Determines whether level 2 headings are treated as dashboard rows or dashboard columns.}

\item{bg}{Background color for \code{bs_theme}}

\item{fg}{Foreground color for \code{bs_theme}}

\item{primary}{Primary color for \code{bs_theme}}

\item{secondary}{Secondary color for code \code{bs_theme}}

\item{version}{Bootstrap version to use with \code{bs_theme}}

\item{fontUrl}{URL where font is hosted}

\item{family}{String. Name of font family when using fontURL}

\item{style}{String. Name of font style. Choose from normal, italic,oblique,
initial, or inherit}

\item{weight}{String. Defines from thin to thick characters.
400 is the same as normal, and 700 is the same as bold. Can also use
normal, bold, bolder, or lighter}

\item{...}{Additional arguments to be passed to \code{flex_dashboard}}
}
\value{
function
}
\description{
Function to create a custom flexdashboard theme for EHA.
}
\section{CSS Font Properties}{

See W3 description of \href{https://www.w3schools.com/cssref/pr_font_font.asp}{font property}
for additional details about customizing fonts.
}

\seealso{
\code{\link[flexdashboard:flex_dashboard]{flexdashboard::flex_dashboard()}}
\code{\link[bslib:bs_theme]{bslib::bs_theme()}}
}
