% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_pptx_template.R
\name{eha_pptx_template}
\alias{eha_pptx_template}
\title{Open EHA Powerpoint templte}
\usage{
eha_pptx_template(
  template = c("classic", "avenir", "predict"),
  aspect = c("16x9", "4x3"),
  outdir = "./",
  recursive = TRUE
)
}
\arguments{
\item{template}{String. Either classic or avenir. Defines the template to be use.}

\item{aspect}{String. Either 16x9 or 4x3. Defines the aspect ratio of the slides}

\item{outdir}{String. Where should files be copied to before opening?}

\item{recursive}{Logical. Should elements of the path other than the last be
created? If true, like the Unix command mkdir -p. See \code{\link[base]{dir.create}}}
}
\value{
pptx
}
\description{
Open avenir, classic, and predict powerpoint templates. Providing no arguments
will open all six templates. Providing either an aspect or a template will
return all the corresponding templates. For example, if classic is provided
to template, then classic 16x9 and classic 4x3 will be opened.
}
\examples{
#open all the templates
\dontrun{eha_pptx_template()}

}
